Moved protobuf files to their own directory so that they're all in the same place (and because .proto files seem to only be able to include other .proto files in the same directory or below, which caused issues for putting, say, interventions.proto in the intervention_model dir when it relied on data.proto in the readers directory)
