+++ src/protobuf/disease.proto
